Update createXDGDotfilesSymlinks to be idempodent

+++ bin/install/common/installs/dotfiles.sh
@@ -45,26 +45,32 @@ function createRegularDotfilesSymlinks() {
 function createXDGDotfilesSymlinks() {
   local THIS_REPO_DIR=$1
 
+  local XDG_CONFIG_DIR=$THIS_REPO_DIR/config
+
   echo "$PFX Creating XDG dotfile symlinks in directory: $HOME/.config"
-  
-  # XDG Base Directory Specification: 
-  # http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html
-  if [ ! -d $HOME/.config ]; then
-    echo "$PFX Creating $HOME/.config"
-    mkdir -p $HOME/.config
-  fi
 
-  CONFIGS=$( find -H "$THIS_REPO_DIR/config" -maxdepth 2 -name '*.symlink' )
-
-  for CONFIG in $CONFIGS; do
-    TARGET="$HOME/.config/$( basename $CONFIG ".symlink" )"
-    if [ -e $TARGET ]; then
-      echo "$PFX ~${TARGET#$HOME} already exists... Skipping."
-    else
-      echo "$PFX Creating $TARGET symlink to $CONFIG"
-      ln -s $CONFIG $TARGET
+  if [ -e $XDG_CONFIG_DIR ]; then
+    # XDG Base Directory Specification: 
+    # http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html
+    if [ ! -d $HOME/.config ]; then
+      echo "$PFX Creating $HOME/.config"
+      mkdir -p $HOME/.config
     fi
-  done
+
+    CONFIGS=$( find -H "$XDG_CONFIG_DIR" -maxdepth 2 -name '*.symlink' )
+
+    for CONFIG in $CONFIGS; do
+      TARGET="$HOME/.config/$( basename $CONFIG ".symlink" )"
+      if [ -e $TARGET ]; then
+        echo "$PFX ~${TARGET#$HOME} already exists... Skipping."
+      else
+        echo "$PFX Creating $TARGET symlink to $CONFIG"
+        ln -s $CONFIG $TARGET
+      fi
+    done
+  else
+    echo "No XDG dotfiles directory in this repo, skipping..."
+  fi
 }
 
 function createDotfilesSymlinks() {
@@ -72,7 +78,6 @@ function createDotfilesSymlinks() {
   createXDGDotfilesSymlinks $REPO_DIR
 
   createRegularDotfilesSymlinks $REPO_LOCAL_DIR
-  createXDGDotfilesSymlinks $REPO_LOCAL_DIR
 }
 
 #
